+++ admin/src/utils/constants.js
@@ -1,0 +1,151 @@
+export const YOUNG_STATUS = {
+  WAITING_VALIDATION: "WAITING_VALIDATION",
+  WAITING_CORRECTION: "WAITING_CORRECTION",
+  VALIDATED: "VALIDATED",
+  REFUSED: "REFUSED",
+  IN_PROGRESS: "IN_PROGRESS",
+  WITHDRAWN: "WITHDRAWN",
+};
+
+export const YOUNG_STATUS_PHASE1 = {
+  AFFECTED: "AFFECTED",
+  WAITING_AFFECTATION: "WAITING_AFFECTATION",
+  CANCEL: "CANCEL",
+  DONE: "DONE",
+  NOT_DONE: "NOT_DONE",
+};
+
+export const YOUNG_STATUS_PHASE2 = {
+  WAITING_REALISATION: "WAITING_REALISATION",
+  VALIDATED: "VALIDATED",
+  IN_PROGRESS: "IN_PROGRESS",
+};
+
+export const YOUNG_PHASE = {
+  INSCRIPTION: "INSCRIPTION",
+  COHESION_STAY: "COHESION_STAY",
+  INTEREST_MISSION: "INTEREST_MISSION",
+};
+
+export const YOUNG_SITUATIONS = {
+  GENERAL_SCHOOL: "GENERAL_SCHOOL",
+  PROFESSIONAL_SCHOOL: "PROFESSIONAL_SCHOOL",
+  AGRICULTURAL_SCHOOL: "AGRICULTURAL_SCHOOL",
+  SPECIALIZED_SCHOOL: "SPECIALIZED_SCHOOL",
+  APPRENTICESHIP: "APPRENTICESHIP",
+  EMPLOYEE: "EMPLOYEE",
+  INDEPENDANT: "INDEPENDANT",
+  SELF_EMPLOYED: "SELF_EMPLOYED",
+  ADAPTED_COMPANY: "ADAPTED_COMPANY",
+  POLE_EMPLOI: "POLE_EMPLOI",
+  MISSION_LOCALE: "MISSION_LOCALE",
+  CAP_EMPLOI: "CAP_EMPLOI",
+  NOTHING: "NOTHING", // @todo find a better key --'
+};
+
+export const YOUNG_STATUS_COLORS = {
+  WAITING_VALIDATION: "#FE7B52",
+  WAITING_CORRECTION: "#FEB951",
+  VALIDATED: "#6CC763",
+  REFUSED: "#F8A9AD",
+  IN_PROGRESS: "#382F79",
+  WITHDRAWN: "#884500",
+};
+
+export const YOUNG_STATUS_PHASE2_COLORS = {
+  WAITING_REALISATION: "#FE7B52",
+  IN_PROGRESS: "#382F79",
+  VALIDATED: "#6CC763",
+};
+
+export const REFERENT_ROLES = {
+  ADMIN: "admin",
+  REFERENT_DEPARTMENT: "referent_department",
+  REFERENT_REGION: "referent_region",
+  RESPONSIBLE: "responsible",
+  SUPERVISOR: "supervisor",
+};
+
+export const MISSION_STATUS = {
+  WAITING_VALIDATION: "WAITING_VALIDATION",
+  WAITING_CORRECTION: "WAITING_CORRECTION",
+  VALIDATED: "VALIDATED",
+  DRAFT: "DRAFT",
+  REFUSED: "REFUSED",
+  CANCEL: "CANCEL",
+  ARCHIVED: "ARCHIVED",
+};
+
+export const MISSION_STATUS_COLORS = {
+  WAITING_VALIDATION: "#FE7B52",
+  WAITING_CORRECTION: "#FEB951",
+  VALIDATED: "#6CC763",
+  DRAFT: "#d9bb71",
+  REFUSED: "#F8A9AD",
+  CANCEL: "#ffa987",
+  ARCHIVED: "#ffb3fb",
+};
+
+export const APPLICATION_STATUS = {
+  WAITING_VALIDATION: "WAITING_VALIDATION",
+  WAITING_ACCEPTATION: "WAITING_ACCEPTATION",
+  VALIDATED: "VALIDATED",
+  REFUSED: "REFUSED",
+  CANCEL: "CANCEL",
+  IN_PROGRESS: "IN_PROGRESS",
+  DONE: "DONE",
+  ABANDON: "ABANDON",
+};
+
+export const APPLICATION_STATUS_COLORS = {
+  WAITING_VALIDATION: "#FE7B52",
+  WAITING_ACCEPTATION: "#FEB951",
+  VALIDATED: "#6CC763",
+  DONE: "#1C7713",
+  REFUSED: "#F8A9AD",
+  CANCEL: "#ffa987",
+  IN_PROGRESS: "#382F79",
+  ABANDON: "#BE3B12",
+};
+
+export const MISSION_PERIOD_DURING_HOLIDAYS = {
+  SUMMER: "SUMMER",
+  AUTUMN: "AUTUMN",
+  DECEMBER: "DECEMBER",
+  WINTER: "WINTER",
+  SPRING: "SPRING",
+};
+
+export const MISSION_PERIOD_DURING_SCHOOL = {
+  EVENING: "EVENING",
+  END_DAY: "END_DAY",
+  WEEKEND: "WEEKEND",
+};
+
+export const MISSION_DOMAINS = {
+  CITIZENSHIP: "CITIZENSHIP",
+  CULTURE: "CULTURE",
+  DEFENSE: "DEFENSE",
+  EDUCATION: "EDUCATION",
+  ENVIRONMENT: "ENVIRONMENT",
+  HEALTH: "HEALTH",
+  SECURITY: "SECURITY",
+  SOLIDARITY: "SOLIDARITY",
+  SPORT: "SPORT",
+};
+
+export const STRUCTURE_STATUS = {
+  WAITING_VALIDATION: "WAITING_VALIDATION",
+  WAITING_CORRECTION: "WAITING_CORRECTION",
+  VALIDATED: "VALIDATED",
+  DRAFT: "DRAFT",
+};
+
+export const STRUCTURE_STATUS_COLORS = {
+  WAITING_VALIDATION: "#FE7B52",
+  WAITING_CORRECTION: "#FEB951",
+  VALIDATED: "#6CC763",
+  DRAFT: "#d9bb71",
+};
+
+export const DEFAULT_STRUCTURE_NAME = "Ma nouvelle Structure";
