+++ admin/src/utils/index.js
@@ -1,0 +1,39 @@
+export * from "./region-and-departments";
+export * from "./constants";
+export * from "./translation";
+export * from "./date";
+
+export const domains = ["Défense et mémoire", "Sécurité", "Solidarité", "Santé", "Éducation", "Culture", "Sport", "Environnement et développement durable", "Citoyenneté"];
+export const status = ["Brouillon", "En attente de validation", "En attente de correction", "Validée", "Refusée", "Annulée", "Archivée"];
+
+export const associationTypes = [
+  "Agrément jeunesse et éducation populaire",
+  "Agrément service civique",
+  "Association complémentaire de l'enseignement public",
+  "Associations d'usagers du système de santé",
+  "Association sportive affiliée à une fédération sportive agréée par l'État",
+  "Agrément des associations de protection de l'environnement",
+  "Association agréée de sécurité civile",
+  "Autre agrément",
+];
+
+export const privateTypes = ["Établissement de santé privé d'intérêt collectif", "Entreprise agréée ESUS", "Autre"];
+export const publicTypes = ["Collectivité territoriale", "Etablissement scolaire", "Etablissement public de santé", "Etablissement public", "Service de l'Etat"];
+export const publicEtatTypes = [
+  "SDIS (Service départemental d'Incendie et de Secours)",
+  "Gendarmerie",
+  "Police",
+  "Armées",
+  "Autre service de l'état",
+  "Autre établissement public",
+];
+
+export const corpsEnUniforme = ["SDIS (Service départemental d'Incendie et de Secours)", "Gendarmerie", "Police", "Armées"];
+
+export const getFilterLabel = (selected, placeholder = "Choisissez un filtre") => {
+  if (Object.keys(selected).length === 0) return placeholder;
+  const translated = Object.keys(selected).map((item) => {
+    return translate(item);
+  });
+  return translated.join(", ");
+};
