+++ app/src/utils/constants.js
@@ -1,0 +1,97 @@
+export const YOUNG_STATUS = {
+  IN_PROGRESS: "IN_PROGRESS",
+  WAITING_VALIDATION: "WAITING_VALIDATION",
+  WAITING_CORRECTION: "WAITING_CORRECTION",
+  VALIDATED: "VALIDATED",
+  REFUSED: "REFUSED",
+};
+
+export const PHASE_STATUS = {
+  IN_PROGRESS: "IN_PROGRESS",
+  IN_COMING: "IN_COMING",
+  VALIDATED: "VALIDATED",
+  CANCEL: "CANCEL",
+  WAITING_AFFECTATION: "WAITING_AFFECTATION",
+};
+export const PHASE_STATUS_COLOR = {
+  VALIDATED: "#6CC763",
+  CANCEL: "#FE7B52",
+};
+
+export const YOUNG_PHASE = {
+  INSCRIPTION: "INSCRIPTION",
+  COHESION_STAY: "COHESION_STAY",
+};
+
+export const APPLICATION_STATUS = {
+  WAITING_VALIDATION: "WAITING_VALIDATION",
+  WAITING_ACCEPTATION: "WAITING_ACCEPTATION",
+  VALIDATED: "VALIDATED",
+  REFUSED: "REFUSED",
+  CANCEL: "CANCEL",
+  IN_PROGRESS: "IN_PROGRESS",
+  DONE: "DONE",
+  ABANDON: "ABANDON",
+};
+
+export const APPLICATION_STATUS_COLORS = {
+  WAITING_VALIDATION: "#FE7B52",
+  WAITING_ACCEPTATION: "#FEB951",
+  VALIDATED: "#6CC763",
+  DONE: "#1C7713",
+  REFUSED: "#F8A9AD",
+  CANCEL: "#ffa987",
+  IN_PROGRESS: "#382F79",
+  ABANDON: "#BE3B12",
+};
+
+export const PERIOD_DURING_HOLIDAYS = {
+  SUMMER: "SUMMER",
+  AUTUMN: "AUTUMN",
+  DECEMBER: "DECEMBER",
+  WINTER: "WINTER",
+  SPRING: "SPRING",
+};
+
+export const PERIOD_DURING_SCHOOL = {
+  EVENING: "EVENING",
+  END_DAY: "END_DAY",
+  WEEKEND: "WEEKEND",
+};
+
+export const MISSION_DOMAINS = {
+  CITIZENSHIP: "CITIZENSHIP",
+  CULTURE: "CULTURE",
+  DEFENSE: "DEFENSE",
+  EDUCATION: "EDUCATION",
+  ENVIRONMENT: "ENVIRONMENT",
+  HEALTH: "HEALTH",
+  SECURITY: "SECURITY",
+  SOLIDARITY: "SOLIDARITY",
+  SPORT: "SPORT",
+};
+
+export const PROFESSIONNAL_PROJECT = {
+  UNIFORM: "UNIFORM",
+  OTHER: "OTHER",
+  UNKNOWN: "UNKNOWN",
+};
+
+export const PROFESSIONNAL_PROJECT_PRECISION = {
+  FIREFIGHTER: "FIREFIGHTER",
+  POLICE: "POLICE",
+  ARMY: "ARMY",
+};
+
+export const PERIOD = {
+  DURING_HOLIDAYS: "DURING_HOLIDAYS",
+  DURING_SCHOOL: "DURING_SCHOOL",
+};
+
+export const TRANSPORT = {
+  PUBLIC_TRANSPORT: "PUBLIC_TRANSPORT",
+  BIKE: "BIKE",
+  MOTOR: "MOTOR",
+  CARPOOLING: "CARPOOLING",
+  OTHER: "OTHER",
+};
