+++ app/src/utils/date.js
@@ -1,0 +1,5 @@
+export const formatStringDate = (date) => {
+  if (!date) return "-";
+  const d = new Date(date);
+  return d.toLocaleDateString("fr-FR", { year: "numeric", month: "long", day: "numeric" });
+};
